refactor(web): consolidate mail mutations into mail queries module
