+++ apps/web/src/components/mail/thread-view.tsx
@@ -18,7 +18,7 @@ import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover
 import { getServerUrl } from '@/lib/api';
 import { getErrorMessage } from '@/lib/errors';
 import { formatDateTime } from '@/lib/format';
-import { useModifyMailMessage, useTrashMailThread, useUntrashMailThread } from '@/lib/mutations/mail';
+import { useModifyMailMessage, useTrashMailThread, useUntrashMailThread } from '@/lib/queries/mail';
 import { mailLabelsQueryOptions, mailThreadQueryOptions } from '@/lib/queries/mail';
 
 type ThreadViewProps = { accountId: MailAccountId; threadId: MailThreadId; onClose: () => void };
